feat: moved request into seperate type
