+++ internal/driver/types.go
@@ -1,0 +1,18 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+//
+// Copyright (C) 2022 Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package driver
+
+type RtspRequest struct {
+	Ip       string `json:"ip"`
+	User     string `json:"user"`
+	Password string `json:"password"`
+	Path     string `json:"path"`
+	Protocol string `json:"protocol"`
+	Id       string `json:"id"`
+	Action   string `json:"action"`
+	Query    string `json:"query"`
+}
